In progress work.
A lot of code needed to reorganized to avoid circular dependencies
