+++ src/fides/api/api/v1/endpoints/saas_config_endpoints.py
@@ -25,8 +25,8 @@ from fides.api.models.sql_models import System  # type: ignore
 from fides.api.oauth.utils import verify_oauth_client
 from fides.api.schemas.connection_configuration.connection_config import (
     SaasConnectionTemplateResponse,
-    SaasConnectionTemplateValues,
 )
+from fides.api.schemas.connection_configuration.saas_config_template_values import SaasConnectionTemplateValues
 from fides.api.schemas.saas.connector_template import ConnectorTemplate
 from fides.api.schemas.saas.saas_config import (
     SaaSConfig,
