+++ src/fides/api/api/v1/endpoints/system.py
@@ -40,8 +40,9 @@ from fides.api.schemas.connection_configuration.connection_config import (
     ConnectionConfigurationResponse,
     CreateConnectionConfigurationWithSecrets,
     SaasConnectionTemplateResponse,
-    SaasConnectionTemplateValues,
 )
+
+from fides.api.schemas.connection_configuration.saas_config_template_values import SaasConnectionTemplateValues
 from fides.api.schemas.system import SystemResponse
 from fides.api.util.api_router import APIRouter
 from fides.api.util.connection_util import (
