+++ src/fides/api/db/seed.py
@@ -29,8 +29,9 @@ from fides.api.models.sql_models import (  # type: ignore[attr-defined]
 from fides.api.oauth.roles import OWNER
 from fides.api.schemas.connection_configuration.connection_config import (
     CreateConnectionConfigurationWithSecrets,
-    SaasConnectionTemplateValues,
 )
+
+from fides.api.schemas.connection_configuration.saas_config_template_values import SaasConnectionTemplateValues
 from fides.api.schemas.dataset import DatasetConfigCtlDataset
 from fides.api.schemas.policy import ActionType, DrpAction
 from fides.api.util.connection_util import patch_connection_configs
