+++ src/fides/api/schemas/connection_configuration/enums/system_type.py
@@ -1,0 +1,8 @@
+from enum import Enum
+
+
+class SystemType(Enum):
+    saas = "saas"
+    database = "database"
+    manual = "manual"
+    email = "email"
