+++ src/fides/api/util/connection_util.py
@@ -30,8 +30,9 @@ from fides.api.schemas.connection_configuration.connection_config import (
     BulkPutConnectionConfiguration,
     ConnectionConfigurationResponse,
     CreateConnectionConfigurationWithSecrets,
-    SaasConnectionTemplateValues,
 )
+
+from fides.api.schemas.connection_configuration.saas_config_template_values import SaasConnectionTemplateValues
 from fides.api.schemas.connection_configuration.connection_secrets_saas import (
     validate_saas_secrets_external_references,
 )
